Add files via upload
Backpropagation in Convolutional Neural Networks (CNNs): It is basically the process of training the network by minimizing the error in predictions through the calculation of gradients. Compute Gradients of Loss w.r.t. Output (Error Signal), Backpropagate Through Fully Connected Layers and Backpropagate Through Convolutional Layers. It involves updating the weights and biases in the network using the chain rule of calculus.